# [NUI.XamlBuild] Remove unused event : Element.DescendantAdded / Removed 