--- a/src/Tizen.NUI/src/public/XamlBinding/Element.cs
+++ b/src/Tizen.NUI/src/public/XamlBinding/Element.cs
@@ -372,11 +372,12 @@ namespace Tizen.NUI.Binding
 
             child.ApplyBindings(skipBindingContext: false, fromBindingContextChanged: true);
 
-            ChildAdded?.Invoke(this, new ElementEventArgs(child));
+            // 2023-11-08 : Just ignore Tizen.NUI don't using logics.
+            // ChildAdded?.Invoke(this, new ElementEventArgs(child));
 
-            OnDescendantAdded(child);
-            foreach (Element element in child.Descendants())
-                OnDescendantAdded(element);
+            // OnDescendantAdded(child);
+            // foreach (Element element in child.Descendants())
+            //     OnDescendantAdded(element);
         }
 
         /// <summary>
@@ -395,11 +396,12 @@ namespace Tizen.NUI.Binding
 
             child.Parent = null;
 
-            ChildRemoved?.Invoke(child, new ElementEventArgs(child));
+            // 2023-11-08 : Just ignore Tizen.NUI don't using logics.
+            // ChildRemoved?.Invoke(child, new ElementEventArgs(child));
 
-            OnDescendantRemoved(child);
-            foreach (Element element in child.Descendants())
-                OnDescendantRemoved(element);
+            // OnDescendantRemoved(child);
+            // foreach (Element element in child.Descendants())
+            //     OnDescendantRemoved(element);
         }
 
         /// <summary>