Those API was useless if we don't need to invoke those events.
And those try to climb-up to parent root every Add / Remove API.

It will spend heavy time if the scene tree is complex.

Those event handerl are internal, and nobody use them.
So just remove codes.