@hinohie 2023-11-08
Hello world @AdunFang .
If then, could I just block the OnDescendantAdded / Removed function?

That API try yo climb up to the scene tree everytime. I think it is useless.