@hinohie 2023-11-08
Performance test when we try to add 4000 View result :

`
    void ElementsTest()
    {
        ++mCreateCount;
        Tizen.Log.Error("NUI.PerfNew", $"Test start!\n");
        DateTime startTime;
        DateTime endTime;

        startTime = DateTime.Now;
        if (elementRootView != null)
        {
            mWindow.Remove(elementRootView);
            elementRootView.DisposeRecursively();
            elementRootView = null;
        }

        this.FlushUpdateMessages();
        endTime = DateTime.Now;
        Tizen.Log.Error("NUI.PerfNew", $"Unparent Spend time : {(endTime - startTime).TotalMilliseconds} ms\n");

        startTime = DateTime.Now;
        elementRootView = new View();
        mWindow.Add(elementRootView);
        View lastView = elementRootView;
        for(int i=0; i<mDepthCount; i++)
        {
            View currentView = new View();
            lastView.Add(currentView);
            lastView = currentView;
        }
        for(int i=0; i<mViewCount; i++)
        {
            View currentView = new View();
            lastView.Add(currentView);
        }
        lastView.SizeWidth = 100;
        lastView.SizeHeight = 100;
        lastView.BackgroundColor = Color.Red;
        this.FlushUpdateMessages();
        endTime = DateTime.Now;
        Tizen.Log.Error("NUI.PerfNew", $"Add Spend time : {(endTime - startTime).TotalMilliseconds} ms\n");

        FullGC();
}
`

* Before patch merged

- depth 1
ERROR: NUI.PerfNew: PerformanceTest.cs: ElementsTest(242) > Average of Add 4000 : 456.174062962963 ms
ERROR: NUI.PerfNew: PerformanceTest.cs: ElementsTest(243) > Average of Remove 4000 : 288.3947851851852 ms
- depth 4000
ERROR: NUI.PerfNew: PerformanceTest.cs: ElementsTest(242) > Average of Add 4000 : 1930.7628037037039 ms
ERROR: NUI.PerfNew: PerformanceTest.cs: ElementsTest(243) > Average of Remove 4000 : 288.90959999999995 ms


* After patch merged

- depth 1
ERROR: NUI.PerfNew: PerformanceTest.cs: ElementsTest(242) > Average of Add 4000 : 440.5272666666667 ms
ERROR: NUI.PerfNew: PerformanceTest.cs: ElementsTest(243) > Average of Remove 4000 : 277.5338222222222 ms
- depth 4000
ERROR: NUI.PerfNew: PerformanceTest.cs: ElementsTest(242) > Average of Add 4000 : 1340.1524925925924 ms
ERROR: NUI.PerfNew: PerformanceTest.cs: ElementsTest(243) > Average of Remove 4000 : 283.14077777777777 ms